cleaned up world rendering in world_editor
